Moves app/config to root config dir
