+++ app/lib/chatbot/controllers/DonorsChooseBotController.js
@@ -30,7 +30,7 @@ var logger = rootRequire('lib/logger');
 var mobilecommons = rootRequire('lib/mobilecommons');
 var shortenLink = rootRequire('lib/bitly');
 var helpers = rootRequire('lib/helpers');
-var connectionOperations = rootRequire('app/config/connectionOperations');
+var connectionOperations = rootRequire('config/connectionOperations');
 var donationModel = require('../models/donorsChooseDonationModel')(connectionOperations);
 
 /**
