+++ app/lib/reportback/test/reportback.test.js
@@ -1,6 +1,6 @@
 var assert = require('assert')
   , reportback = require('../')
-  , connectionOperations = rootRequire('app/config/connectionOperations')
+  , connectionOperations = rootRequire('config/connectionOperations')
   , model = require('../reportbackModel')(connectionOperations)
   , emitter = rootRequire('app/eventEmitter')
   , dscontentapi = rootRequire('lib/phoenix')()
