+++ server.js
@@ -28,9 +28,9 @@ phoenix.userLogin(
  */
 app = express();
 
-var appConfig = require('./app/config')();
+var appConfig = require('./config')();
 var router = require('./app/router');
-var smsConfigsLoader = require('./app/config/smsConfigsLoader');
+var smsConfigsLoader = require('./config/smsConfigsLoader');
 
 // Retrieves all SMS config files before starting server.
 smsConfigsLoader(function() {
