+++ test/CreateGameFromMobile.js
@@ -1,7 +1,7 @@
 var assert = require('assert')
   , express = require('express')
   , emitter = rootRequire('app/eventEmitter')
-  , connectionOperations = rootRequire('app/config/connectionOperations')
+  , connectionOperations = rootRequire('config/connectionOperations')
   , gameCreateConfigModel = rootRequire('app/lib/sms-games/models/sgGameCreateConfig')(connectionOperations)
   , gameModel = rootRequire('app/lib/sms-games/models/sgCompetitiveStory')(connectionOperations)
   , SGCreateFromMobileController = rootRequire('app/lib/sms-games/controllers/SGCreateFromMobileController')
